% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child-runtime.R
\name{child_runtime_code}
\alias{child_runtime_code}
\title{Generate the R code to inject into the child process}
\usage{
child_runtime_code()
}
\value{
A single character string of R code
}
\description{
Returns a character string of R code that, when evaluated in the child
process, sets up the IPC connection and defines \code{.securer_call_tool()}.
}
\keyword{internal}
