% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{ipc_create_server}
\alias{ipc_create_server}
\title{Create a Unix domain socket server}
\usage{
ipc_create_server(path)
}
\arguments{
\item{path}{Character string, the file path for the socket}
}
\value{
A server connection object from processx
}
\description{
Create a Unix domain socket server
}
\keyword{internal}
