% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-registry.R
\docType{data}
\name{type_check_map}
\alias{type_check_map}
\title{Map of type annotation strings to R type-checking functions}
\format{
An object of class \code{character} of length 6.
}
\usage{
type_check_map
}
\description{
Map of type annotation strings to R type-checking functions
}
\keyword{internal}
